Update draft-ietf-teas-5g-ns-ip-mpls.md
To address Jie's comment on 28th Aug 2024, in response to the second wglc. The comment was:

"2404    7.2.2.  Scheme 2: TE Paths with Fixed Bandwidth Reservations
 
2410       slice.  Note that the "reservations" would be in the mind of the
2411       transport controller - it is not necessary (or indeed possible for
2412       SR-TE) to reserve bandwidth at the network layer.
 
[Jie] With some mechanisms resource reservation can also be done on the network devices. If section 7.2.2 only talks about bandwidth reservation in the mind of the controller, it is suggested to make this explicit in the title of this section.
 
As for segment routing with resource reservation at the network layer, this document may consider to add a reference to draft-ietf-spring-resource-aware-segments."

I changed the wording from "would" to "could".